refactor: rename AuthService.ts to authService.ts for consistent naming
